requirements.txt
sqlalchemy
file-magic
curses-utils
